pyproject.toml
[project]
name = "git-ci-gym"
version = "0.1.0"
description = "RL environment for semantic code merging and automated CI repair"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn",
    "pydantic>=2.0",
    "fastmcp",
    "pytest",
    "openai",
]

[project.optional-dependencies]
dev = [
    "python-dotenv",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = []
build-backend = "build_backend"
backend-path = ["."]
